backend/requirements.txt
alembic==1.13.3
fastapi==0.115.3
psycopg2-binary==2.9.10
python-jose==3.3.0
python-dotenv==1.0.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0
SQLAlchemy==2.0.36
uvicorn==0.32.0
web3==7.4.0
bcrypt==4.0.1
passlib==1.7.4
email_validator==2.2.0
cryptography==43.0.3